+++ src/router.js
@@ -4,6 +4,7 @@ import Router from 'vue-router';
 // Import components to route to
 import Home from '@/components/Home';
 import Dashboard from '@/components/Dashboard';
+import ResourcePacks from '@/components/ResourcePacks';
 
 Vue.use(Router);
 
@@ -19,6 +20,11 @@ export default new Router({
             path: '/dashboard',
             name: 'Dashboard',
             component: Dashboard
+        },
+        {
+            path: '/packs',
+            name: 'ResourcePacks',
+            component: ResourcePacks
         }
     ]
 })